+++ src/video/webgl/compositors/quad_compositor.js
@@ -90,6 +90,9 @@ var V_ARRAY = [
         if (pixels === null || typeof pixels.byteLength !== "undefined") {
             // if pixels is undefined, or if it's Uint8Array/Float32Array TypedArray
             gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, w, h, 0, gl.RGBA, gl.UNSIGNED_BYTE, pixels, 0);
+        } else if (pixels instanceof OffscreenCanvas) {
+            // convert to ImageBitmap first (else Safari 16.4 and higher will throw an TypeError exception)
+            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, pixels.transferToImageBitmap());
         } else {
             gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, pixels);
         }